This kinda works for `.gui`
It pollutes the runfiles dir, which persists between runs and creates confusion
(i.e. mkdir and ln fail due to files already existing). Setting up in a temp
dir might be a suitable solution, but would need to be careful _not_ to do that
when the wrapper is used during the build.

Also, using `basename` for the libraries isn't going to be good enough if there
is any hierarchy there. Need to come up with a solution that preserves the
hierarchy.